renamed core export package to exporter package for consistency
