+++ core/src/main/kotlin/com/ritense/valtimo/autoconfiguration/ExportAutoConfiguration.kt
@@ -18,8 +18,8 @@
 package com.ritense.valtimo.autoconfiguration
 
 import com.ritense.valtimo.camunda.service.CamundaRepositoryService
-import com.ritense.valtimo.export.DecisionDefinitionExporter
-import com.ritense.valtimo.export.ProcessDefinitionExporter
+import com.ritense.valtimo.exporter.DecisionDefinitionExporter
+import com.ritense.valtimo.exporter.ProcessDefinitionExporter
 import org.camunda.bpm.engine.RepositoryService
 import org.springframework.boot.autoconfigure.AutoConfiguration
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
