+++ core/src/test/kotlin/com/ritense/valtimo/exporter/DecisionDefinitionExporterIntTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.ritense.valtimo.export
+package com.ritense.valtimo.exporter
 
 import com.ritense.authorization.AuthorizationContext.Companion.runWithoutAuthorization
 import com.ritense.exporter.request.DecisionDefinitionExportRequest
